Capture Put Request Working
Got the PUT request for a Pokemon capture working. This request accepts an email parameter as the primary key, and a json object in the body with the key-value pair of name and the name of the Pokemon. The header type for the name must be of content type application/json.